# service/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  ApplicationListResponse:
    properties:
      apply_list:
        items:
          $ref: '#/definitions/ApplyUserItem'
        type: array
      count:
        type: integer
    type: object
  ApplyUserItem:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  CreateGroupRequest:
    properties:
      group_name:
        type: string
      user_list:
        items:
          type: integer
        type: array
    type: object
...
  CreateRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
  DeleteTrashbinRequest:
    properties:
      type:
        type: integer
    type: object
...
  Group:
    properties:
      id:
        type: integer
      name:
        type: string
      user_count:
        type: integer
    type: object
  GroupListResponse:
    properties:
      count:
        type: integer
      groups:
        items:
          $ref: '#/definitions/Group'
        type: array
    type: object
...
  ListRequest:
    properties:
      group:
        type: integer
      team:
        type: integer
    type: object
  ListResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/User'
        type: array
    type: object
...
  RemoveTrashbinRequest:
    properties:
      children_position_index:
        type: integer
      fatherId:
        type: integer
      is_father_project:
        type: boolean
      type:
        type: integer
    type: object
  Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  Status:
    properties:
      avatar:
        type: string
      comment_count:
        type: integer
      content:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      liked:
        type: boolean
      time:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
  StatusListResponse:
    properties:
      count:
        type: integer
      stauts:
        items:
          $ref: '#/definitions/Status'
        type: array
    type: object
...
  UpdateFolderRequest:
    properties:
      name:
        type: string
    type: object
...
  User:
    properties:
      avatar:
        type: string
      email:
        type: string
      group:
        type: integer
      id:
        type: integer
      name:
        type: string
      real_name:
        type: string
      role:
        type: integer
      team:
        type: integer
    type: object
...
host: work.test.muxi-tech.xyz
info:
  contact: {}
  description: The gateway of muxi-workbench
  title: muxi-workbench-gateway
  version: "1.0"
paths:
...
  /folder/filefolder/{id}:
    delete:
      consumes:
      - application/json
      description: 删除文件文件夹
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: file_folder_id
        in: path
        name: id
        required: true
        type: integer
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: delete a file folder api
      tags:
      - project
    put:
      consumes:
      - application/json
      description: 修改文件夹名字
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 此文件夹 id
        in: path
        name: id
        required: true
        type: integer
      - description: update_folder_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/UpdateFolderRequest'
      - description: 此文件夹所属项目 id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: update file folder's name api
      tags:
      - project
...
  /status:
    get:
      consumes:
      - application/json
      description: 拉取进度列表
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: last_id
        in: query
        name: last_id
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: group
        in: query
        name: group
        type: integer
      - description: uid
        in: query
        name: uid
        type: integer
      - description: team
        in: query
        name: team
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/StatusListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: list status api
      tags:
      - status
    post:
      consumes:
      - application/json
      description: 创建 status
      parameters:
      - description: create_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/CreateRequest'
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: create status api
      tags:
      - status
...
  /team/application/list:
    get:
      consumes:
      - application/json
      description: 拉取 application 列表
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page 从 0 开始计数， 如果传入非负数或者不传值则不分页
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApplicationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: list application api
      tags:
      - application
  /team/group:
    get:
      consumes:
      - application/json
      description: 拉取 group 列表
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page 从 0 开始计数， 如果传入非负数或者不传值则不分页
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GroupListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: list group api
      tags:
      - group
    post:
      consumes:
      - application/json
      description: 创建 group
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: create_group_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: create group api
      tags:
      - group
...
  /trashbin/{id}:
    delete:
      consumes:
      - application/json
      description: 删除回收站内文件,其中 type：0-project 1-doc 2-file 3-doc folder 4-file folder
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: file_id
        in: path
        name: id
        required: true
        type: integer
      - description: delete_trashbin_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/DeleteTrashbinRequest'
      - description: project_id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: delete a trashbin api
      tags:
      - project
    put:
      consumes:
      - application/json
      description: 恢复回收站资源(type：0-project 1-doc 2-file 3-doc_folder 4-file_folder)
      parameters:
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 回收站中文件的 id
        in: path
        name: id
        required: true
        type: integer
      - description: remove_trashbin_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/RemoveTrashbinRequest'
      - description: 此回收站对应的项目 id
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: update trashbin api
      tags:
      - project
...
  /user/list:
    get:
      consumes:
      - application/json
      description: 通过 group 和 team 获取 userlist
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: token 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: get_user_list_request
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: get user_list api
      tags:
      - user
...
swagger: "2.0"
...